Move dumps handler into main dir
